Placeholder dummy need to mimic original element better
The placing of a dummy element that's suppose to fill the space of the original after its been sticky'd was fairly naive in it's assumptions with regards to the positioning of the original element. I don't believe I've covered everything that may affect the original element's layout, but it should cover most cases now, unlike before, where it would break layout if something as simple as `float` was declared on the original element.